Move CCSDTQ functionality into solver_rewrite.
